💥 Add UnknownDeviceException
